+++ src/simmate/calculators/vasp/workflows/relaxation/mit.py
@@ -14,6 +14,6 @@ workflow = s3task_to_workflow(
     project_name="Simmate-Relaxation",
     s3task=MITRelaxationTask,
     calculation_table=MITRelaxationResults,
-    register_kwargs=["prefect_flow_run_id", "structure", "source"],
+    register_kwargs=["structure", "source"],
     description_doc_short="uses MIT Project settings",
 )