+++ src/simmate/calculators/vasp/workflows/relaxation/neb_endpoint.py
@@ -14,6 +14,6 @@ workflow = s3task_to_workflow(
     project_name="Simmate-Relaxation",
     s3task=NEBEndpointRelaxationTask,
     calculation_table=NEBEndpointRelaxationResults,
-    register_kwargs=["prefect_flow_run_id", "structure", "source"],
+    register_kwargs=["structure", "source"],
     description_doc_short="uses Materials Project settings and meant for endpoint supercells",
 )