+++ src/simmate/calculators/vasp/workflows/static_energy/quality_04.py
@@ -14,6 +14,6 @@ workflow = s3task_to_workflow(
     project_name="Simmate-Energy",
     s3task=Quality04EnergyTask,
     calculation_table=Quality04StaticEnergyResults,
-    register_kwargs=["prefect_flow_run_id", "structure", "source"],
+    register_kwargs=["structure", "source"],
     description_doc_short="low-quality settings meant for highly unreasonable structures",
 )