+++ src/simmate/configuration/example_project/example_app/workflows.py
@@ -31,7 +31,7 @@ example_workflow = s3task_to_workflow(
     # this sets what should be saved to the database BEFORE the calculation
     # is actually started. This helps you search your database for calculations
     # that have not completed yet.
-    register_kwargs=["prefect_flow_run_id", "structure", "source"],
+    register_kwargs=["structure", "source"],
     # Quick description that will be used in the website interface
     description_doc_short="This is my new fancy workflow!",
 )