+++ src/simmate/website/workflows/test/test_workflows_views.py
@@ -122,4 +122,7 @@ def test_workflow_submit_view(client, sample_structures, mocker):
         source=None,
         copy_previous_directory=False,
         wait_for_run=False,
+        # parameters not deserialized yet so these will still be present
+        command="",
+        directory="",
     )